Removed Package-lock.json from repo and refactored get questions
