+++ .gitignore
@@ -1,6 +1,4 @@
 node_modules/
 coverage/
 .nyc_output/
-npm-debug.log
-.DS_Store
-/*.env
+package-lock.json
