--- api_build/package.json
@@ -1,61 +1,0 @@
-{
-  "name": "api",
-  "version": "0.0.0",
-  "description": "APIs for StackOverflow-lite web application",
-  "main": "app.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/TheoOkafor/StackOverflow-lite.git"
-  },
-  "keywords": [
-    "Questions",
-    "Answers"
-  ],
-  "author": "Theo Okafor",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/TheoOkafor/StackOverflow-lite/issues"
-  },
-  "homepage": "https://github.com/TheoOkafor/StackOverflow-lite#readme",
-  "scripts": {
-    "start": "node app.js",
-    "devstart": "nodemon app.js",
-    "test": "mocha --exit test/test.js",
-    "cover": "istanbul cover _mocha --exit",
-    "babel-app": "babel --presets es2015  app.js --out-file ../app.js",
-    "babel-test": "babel --presets es2015  test/test.js --out-file ../test/test.js",
-    "babel-data": "babel --presets es2015  data/data.js --out-file ../data/data.js",
-    "babel-users": "babel --presets es2015  routes/users.js --out-file ../routes/users.js",
-    "babel-questions": "babel --presets es2015  routes/questions.js --out-file ../routes/questions.js",
-    "coveralls": "npm run cover -- --report lcovonly && cat ./coverage/lcov.info | coveralls --exit"
-  },
-  "dependencies": {
-    "body-parser": "^1.18.3",
-    "config": "^2.0.1",
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.6.9",
-    "express": "~4.16.0",
-    "http-errors": "~1.6.2",
-    "morgan": "~1.9.0"
-  },
-  "devDependencies": {
-    "babel": "^6.23.0",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-plugin-check-es2015-constants": "^6.22.0",
-    "babel-plugin-transform-es2015-arrow-functions": "^6.22.0",
-    "babel-plugin-transform-es2015-block-scoping": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "chai": "^4.1.2",
-    "chai-http": "^4.0.0",
-    "coveralls": "^3.0.2",
-    "istanbul": "^0.4.5",
-    "jscover": "^1.0.0",
-    "mocha": "^5.2.0",
-    "mocha-lcov-reporter": "^1.3.0",
-    "nodemon": "^1.18.3"
-  },
-  "engines": {
-    "node": "8.11.3"
-  }
-}
