+++ package.json
@@ -18,9 +18,9 @@
   },
   "homepage": "https://github.com/TheoOkafor/StackOverflow-lite#readme",
   "scripts": {
-    "start": "node app.js",
-    "devstart": "nodemon app.js",
-    "test": "nyc mocha --timeout=3000 --exit",
+    "start": "babel-node app.js",
+    "devstart": "nodemon app.js --exec babel-node",
+    "test": "nyc mocha --timeout=3000 --require babel-register test --exec",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "dependencies": {
@@ -30,25 +30,32 @@
     "debug": "~2.6.9",
     "express": "~4.16.0",
     "http-errors": "~1.6.2",
-    "morgan": "~1.9.0"
+    "morgan": "~1.9.0",
+    "eslint": "^5.3.0",
+    "eslint-config-airbnb": "^17.1.0"
   },
   "devDependencies": {
     "babel": "^6.23.0",
     "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-plugin-check-es2015-constants": "^6.22.0",
-    "babel-plugin-transform-es2015-arrow-functions": "^6.22.0",
-    "babel-plugin-transform-es2015-block-scoping": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
+    "babel-eslint": "^8.2.6",
+    "babel-preset-env": "^1.7.0",
+    "babel-register": "^6.26.0",
     "chai": "^4.1.2",
     "chai-http": "^4.0.0",
     "coveralls": "^3.0.2",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-config-prettier": "^3.0.1",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-prettier": "^2.6.2",
+    "eslint-plugin-react": "^7.11.1",
     "istanbul": "^0.4.5",
     "jscover": "^1.0.0",
     "mocha": "^5.2.0",
     "mocha-lcov-reporter": "^1.3.0",
     "nodemon": "^1.18.3",
-    "nyc": "^12.0.2"
+    "nyc": "^12.0.2",
+    "prettier": "^1.14.2"
   },
   "engines": {
     "node": "8.11.3"
