+++ test/test.js
@@ -1,67 +1,22 @@
-'use strict';
-
 // Require the dev-dependencies
+import chai from 'chai';
+import chaiHttp from 'chai-http';
+import {data} from './data/data';
+import {app} from './app';
 
-var _chai = require('chai');
+const questions = data.questions;
 
-var _chai2 = _interopRequireDefault(_chai);
-
-var _chaiHttp = require('chai-http');
-
-var _chaiHttp2 = _interopRequireDefault(_chaiHttp);
-
-var _data = require('../data/data');
-
-var _app = require('../app');
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-var questions = _data.data.questions;
-
-_chai2.default.use(_chaiHttp2.default);
+chai.use(chaiHttp);
 // Parent block for QUESTIONS
 describe('Questions', function () {
   describe('GET QUESTIONS', function () {
     // Test the /GET route
-    describe('/GET questions', function () {
-      it('it should GET all the questions', function (done) {
-        _chai2.default.request(_app.app).get('/').end(function (err, res) {
-          _chai2.default.expect(res).to.have.status(200);
-          _chai2.default.expect(res.body).be.a('array');
-          _chai2.default.expect(res.body).to.have.lengthOf(4);
-          done(err);
-        });
-      });
-    });
-
-    describe('/GET questions', function () {
-      it('it should GET all the questions', function (done) {
-        _chai2.default.request(_app.app).get('/questions').end(function (err, res) {
-          _chai2.default.expect(res).to.have.status(200);
-          _chai2.default.expect(res.body).be.a('array');
-          _chai2.default.expect(res.body).to.have.lengthOf(4);
-          done(err);
-        });
-      });
-    });
-
     describe('/GET v1/questions', function () {
       it('it should GET all the questions', function (done) {
-        _chai2.default.request(_app.app).get('/v1/questions').end(function (err, res) {
-          _chai2.default.expect(res).to.have.status(200);
-          _chai2.default.expect(res.body).be.a('array');
-          _chai2.default.expect(res.body).to.have.lengthOf(4);
-          done(err);
-        });
-      });
-    });
-
-    describe('/GET /v1', function () {
-      it('it should GET all the questions', function (done) {
-        _chai2.default.request(_app.app).get('/v1').end(function (err, res) {
-          _chai2.default.expect(200);
-          _chai2.default.expect(res.body).be.a('array');
-          _chai2.default.expect(res.body).to.have.lengthOf(4);
+        chai.request(app).get('/v1/questions').end(function (err, res) {
+          chai.expect(res).to.have.status(200);
+          chai.expect(res.body).be.a('array');
+          chai.expect(res.body).to.have.lengthOf(4);
           done(err);
         });
       });
@@ -70,10 +25,10 @@ describe('Questions', function () {
     // GET A QUESTION FROM QUESTIONS TEST
     describe('/GET /v1/questions/2', function () {
       it('it should GET the question with id = 2', function (done) {
-        _chai2.default.request(_app.app).get('/v1/questions/2').end(function (err, res) {
-          _chai2.default.expect(res).to.have.status(200);
-          _chai2.default.expect(res.body).be.a('object');
-          _chai2.default.expect(res.body.id).to.equal(2);
+        chai.request(app).get('/v1/questions/2').end(function (err, res) {
+          chai.expect(res).to.have.status(200);
+          chai.expect(res.body).be.a('object');
+          chai.expect(res.body.id).to.equal(2);
           done(err);
         });
       });
@@ -81,10 +36,10 @@ describe('Questions', function () {
 
     describe('/GET /v1/questions/100', function () {
       it('it should return \'Question 100 Not Found\'', function (done) {
-        _chai2.default.request(_app.app).get('/v1/questions/100').end(function (err, res) {
-          _chai2.default.expect(res).to.have.status(404);
-          _chai2.default.expect(res.body).be.a('string');
-          _chai2.default.expect(res.body).to.equal('Question 100 Not Found');
+        chai.request(app).get('/v1/questions/100').end(function (err, res) {
+          chai.expect(res).to.have.status(404);
+          chai.expect(res.body).be.a('string');
+          chai.expect(res.body).to.equal('Question 100 Not Found');
           done(err);
         });
       });
@@ -95,14 +50,15 @@ describe('Questions', function () {
   describe('POST QUESTIONS', function () {
     describe('/POST /v1/questions', function () {
       it('it should NOT POST Question if the TITLE is Not provided', function (done) {
-        var question = {
+        const question = {
           body: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
           username: 'TheoOkafor'
         };
-        _chai2.default.request(_app.app).post('/v1/questions').send(question).end(function (err, res) {
-          _chai2.default.expect(res).to.have.status(400);
-          _chai2.default.expect(res.body).to.be.a('object');
-          _chai2.default.expect(res.body).to.have.property('message');
+        chai.request(app).post('/v1/questions').send(question).end(function (err, res) {
+          chai.expect(res).to.have.status(400);
+          chai.expect(res.body).to.be.a('object');
+          chai.expect(res.body).to.have.property('message');
+          //message value here
           done(err);
         });
       });
@@ -110,16 +66,16 @@ describe('Questions', function () {
 
     describe('/POST /v1/questions', function () {
       it('it should POST Question if all the required fields are provided', function (done) {
-        var question = {
+        const question = {
           title: 'Why do people hate reading?',
           body: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
           username: 'TheoOkafor'
         };
-        _chai2.default.request(_app.app).post('/v1/questions').send(question).end(function (err, res) {
-          _chai2.default.expect(res).to.have.status(200);
-          _chai2.default.expect(res.body).to.be.a('object');
-          _chai2.default.expect(res.body).to.have.property('message');
-          _chai2.default.expect(res.body).to.have.property('location');
+        chai.request(app).post('/v1/questions').send(question).end(function (err, res) {
+          chai.expect(res).to.have.status(201);
+          chai.expect(res.body).to.be.a('object');
+          chai.expect(res.body).to.have.property('message');
+          chai.expect(res.body).to.have.property('location');
           done(err);
         });
       });
@@ -127,11 +83,11 @@ describe('Questions', function () {
 
     describe('/GET /v1/questions/' + questions[questions.length - 1].id, function () {
       it('it should return GET the question ' + questions[questions.length - 1].id + ' that was just POSTed', function (done) {
-        _chai2.default.request(_app.app).get('/v1/questions/' + questions[questions.length - 1].id).end(function (err, res) {
-          _chai2.default.expect(res).to.have.status(200);
-          _chai2.default.expect(res.body).to.be.a('object');
-          _chai2.default.expect(res.body).to.have.property('id');
-          _chai2.default.expect(res.body.title).to.equal('Why do people hate reading?');
+        chai.request(app).get('/v1/questions/' + questions[questions.length - 1].id).end(function (err, res) {
+          chai.expect(res).to.have.status(200);
+          chai.expect(res.body).to.be.a('object');
+          chai.expect(res.body).to.have.property('id');
+          chai.expect(res.body.title).to.equal('Why do people hate reading?');
           done(err);
         });
       });
@@ -143,25 +99,15 @@ describe('POST Answers', function () {
   // TESTS FOR ANSWERS
   describe('/POST /v1/questions/16/answers', function () {
     it('it should NOT POST Answer if question does not exist', function (done) {
-      var answer = {
+      const answer = {
         body: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
         username: 'TheoOkafor'
       };
-      _chai2.default.request(_app.app).post('/v1/questions/16/answers').send(answer).end(function (err, res) {
-        _chai2.default.expect(res).to.have.status(404);
-        _chai2.default.expect(res.body).to.be.a('object');
-        _chai2.default.expect(res.body).to.have.property('message');
-        _chai2.default.expect(res.body.message).to.equal('Question 16 Not Found');
-        done(err);
-      });
-    });
-  });
-  describe('/POST /v1/questions/1/answers', function () {
-    it('it should return error for a request server cannot handle', function (done) {
-      var answer = ['This is a test', 'TheoOkafor'];
-      _chai2.default.request(_app.app).post('/v1/questions/1/answers').send(answer).end(function (err, res) {
-        _chai2.default.expect(res).to.have.status(500);
-        _chai2.default.expect(res.body).to.be.a('string');
+      chai.request(app).post('/v1/questions/16/answers').send(answer).end(function (err, res) {
+        chai.expect(res).to.have.status(404);
+        chai.expect(res.body).to.be.a('object');
+        chai.expect(res.body).to.have.property('message');
+        chai.expect(res.body.message).to.equal('Question 16 Not Found');
         done(err);
       });
     });
@@ -169,14 +115,14 @@ describe('POST Answers', function () {
 
   describe('/POST /v1/questions/1/answers', function () {
     it('it should NOT POST Answer if the BODY is Not provided', function (done) {
-      var answer = {
+      const answer = {
         username: 'TheoOkafor'
       };
-      _chai2.default.request(_app.app).post('/v1/questions/1/answers').send(answer).end(function (err, res) {
-        _chai2.default.expect(res).to.have.status(400);
-        _chai2.default.expect(res.body).to.be.a('object');
-        _chai2.default.expect(res.body).to.have.property('message');
-        _chai2.default.expect(res.body.message).to.equal('Bad Request. Answer must have a body.');
+      chai.request(app).post('/v1/questions/1/answers').send(answer).end(function (err, res) {
+        chai.expect(res).to.have.status(400);
+        chai.expect(res.body).to.be.a('object');
+        chai.expect(res.body).to.have.property('message');
+        chai.expect(res.body.message).to.equal('Bad Request. Answer must have a body.');
         done(err);
       });
     });
@@ -184,15 +130,15 @@ describe('POST Answers', function () {
 
   describe('/POST /v1/questions/1/answers', function () {
     it('it should POST answer if all the required fields are provided', function (done) {
-      var answer = {
+      const answer = {
         body: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
         username: 'TheoOkafor'
       };
-      _chai2.default.request(_app.app).post('/v1/questions/1/answers').send(answer).end(function (err, res) {
-        _chai2.default.expect(res).to.have.status(200);
-        _chai2.default.expect(res.body).to.be.a('object');
-        _chai2.default.expect(res.body).to.have.property('message');
-        _chai2.default.expect(res.body).to.have.property('question');
+      chai.request(app).post('/v1/questions/1/answers').send(answer).end(function (err, res) {
+        chai.expect(res).to.have.status(201);
+        chai.expect(res.body).to.be.a('object');
+        chai.expect(res.body).to.have.property('message');
+        chai.expect(res.body).to.have.property('question');
         done(err);
       });
     });
